perf(cli): classify directly; keep the spawns that test a process
Closes #1161. `affected-classifier.spec.ts` had 40 tests and every one
spawned the built CLI. Measured here: the `gjsify affected classifier`
describe took 12.81 s against a 30 s per-`describe()` budget — a 2.6x
margin, matching the 11.68 s / 12.30 s the issue measured on two other
Linux hosts, so it is a property of the suite and not of one host's mood.
A macOS runner only has to be 2.6x slower per spawn to blow it, and on CI
macOS (SIP, Gatekeeper, shared tenancy) 2-5x is ordinary. The failure was
therefore a BUDGET, never an assertion, on the platform whose suites run
on `main` and the nightly rather than on PRs.

THIS OVERRIDES A DOCUMENTED DECISION, so it says why. The spec's header
justified the spawning: it kept `classifyAndExpand` unexported, "a
private detail of the command". The privacy is kept and the spawns are
not — the classifier and its tables move to
`commands/affected-classify.ts`, which is NOT in the package's `exports`
map, so the spec reaches it by relative path and the public surface is
byte-unchanged. That header moved with the code it justifies rather than
being deleted.

What the spawn bought beyond privacy was the argv/stdin/JSON round trip,
and 7 tests still buy it: both emit branches (stdout and the
GITHUB_OUTPUT appendFileSync path) go through the real CLI, and the
word-split tests spawn a real `sh`, which is the only thing that can tell
a token from a token-with-quotes-attached. Those genuinely cross a
process boundary. The other 33 were a pure function tested through one.

The fixture is unchanged — a real temp monorepo walked by the real
`discoverWorkspaces` — so no case swaps a walked tree for a hand-built
one.

Measured after: classifier describe 12.81 s -> 2.19 s, whole CLI suite
14.97 s -> 3.90 s, and the same 2081 tests pass. The margin goes from
~2.6x to ~14x, which is what stops it being a platform property. (Not the
~50 ms the issue projected: the 7 spawns remain, and discovery still
walks the fixture once per case.)
